requirements.txt
ollama>=0.3.0
python-slugify>=8.0.4